fix&&chorn：<碰撞系统>修正一些错误，规范注释与命名

+++ Unity/Assets/Model/NKGMOBA/Battle/Box2DSharp/CollisionRelations/B2S_CollisionRelationSupport.cs
@@ -14,7 +14,7 @@ namespace ETModel
     public class B2S_CollisionsRelationSupport
     {
         [LabelText("此数据载体ID")]
-        public int SupportId;
+        public long SupportId;
 
         [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)]
         public Dictionary<long, B2S_CollisionInstance> B2S_CollisionsRelationDic = new Dictionary<long, B2S_CollisionInstance>();
